% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baggedClusterETS.R
\name{baggedClusterETS}
\alias{baggedClusterETS}
\title{Perform BaggedClusterETS method}
\usage{
baggedClusterETS(y, cores = detectCores() - 1, nclusters = 5,
  distance = "EUCL", silhouette = F, h_pseudo = NULL,
  boot_samples = 1000, ...)
}
\arguments{
\item{y}{a time series}

\item{cores}{(optional) number of cores used to do the calculations}

\item{nclusters}{(optional) number of clusters used. If not selected it uses the Silhouette method to define it}

\item{distance}{distance function used to create the clusters. Default is the Euclidian distance but one can choose among many from the TSclust package}

\item{silhouette}{If TRUE the method performs the Silhouette method and ignores the nclusters selection}

\item{h_pseudo}{number of observations to the validation set that will be used to select the best series}

\item{boot_samples}{number of bootstrap series generated to perform the cluster phase}

\item{...}{other parameters that affect the ETS estimation}
}
\value{
a list with several elements such as the generated bootstrapped series, fitted method to each of the bootstrapped series and so on
}
\description{
BaggedClusterETS method
}
\examples{
\dontrun{
model <- baggedCLusterETS(gas)
plot(forecast(model))
}
}
